ISSUE-285: assertFileSystemWrites
