+++ tests/Domain/App/BuildAppFilesTestCase.php
@@ -7,14 +7,14 @@ namespace App\Tests\Domain\App;
 use App\Infrastructure\CQRS\Bus\CommandBus;
 use App\Infrastructure\CQRS\Bus\DomainCommand;
 use App\Tests\ContainerTestCase;
-use App\Tests\Infrastructure\FileSystem\ProvideFileSystemWriteAssertion;
 use App\Tests\ProvideTestData;
 use League\Flysystem\FilesystemOperator;
+use Spatie\Snapshots\MatchesSnapshots;
 
 abstract class BuildAppFilesTestCase extends ContainerTestCase
 {
     use ProvideTestData;
-    use ProvideFileSystemWriteAssertion;
+    use MatchesSnapshots;
 
     abstract protected function getDomainCommand(): DomainCommand;
 
@@ -29,6 +29,21 @@ abstract class BuildAppFilesTestCase extends ContainerTestCase
         /** @var \App\Tests\Infrastructure\FileSystem\SpyFileSystem $fileSystem */
         $fileSystem = $this->getContainer()->get(FilesystemOperator::class);
         $this->assertFileSystemWrites($fileSystem->getWrites());
+    }
+
+    public function assertFileSystemWrites(array $writes): void
+    {
+        foreach ($writes as $location => $content) {
+            if (str_ends_with($location, '.json')) {
+                $this->assertMatchesJsonSnapshot($content);
+                continue;
+            }
+            if (str_ends_with($location, '.html')) {
+                $this->assertMatchesHtmlSnapshot($content);
+                continue;
+            }
+            $this->assertMatchesTextSnapshot($content);
+        }
     }
 
     #[\Override]
